Move error to relayer and update relayer baseConfig
