+++ universal-login-relayer/src/config/config.rinkeby.ts
@@ -8,8 +8,6 @@ export const getConfig = () => ({
   ...baseConfig,
   network: 'rinkeby',
   jsonRpcUrl: `https://eth-rinkeby.alchemyapi.io/v2/${getEnv('ALCHEMY_API_KEY', '')}`,
-  transporterEmail: getEnv('TRANSPORTER_EMAIL', ''),
-  transporterEmailPassword: getEnv('TRANSPORTER_EMAIL_PASSWORD', ''),
   ensAddress: '0x00000000000C2E074eC69A0dFb2997BA6C7d2e1e',
   ensRegistrars: ['unilogin.test'],
   ensRegistrar: '0x1BA9B86331DDcD63c2AcA7a98886F7e7C2788cD4',
